[ #166664731] updating readme file
