+++ UI/css/user_welcome.css
@@ -203,8 +203,11 @@ div.div_rule{
     max-width: 100%;
   }
 .text{
-  padding:2%
+  padding:2%;
+  color:black;
+  font-size:20px;
 }
+
   .car_card button {
     background: gray;
     border: 0;
@@ -273,8 +276,8 @@ div.div_rule{
 
 input[type=text], input[type=password], select,textarea,input[type=file] {
   width: 100%;
-  padding: 10px 10px;
-  margin: 8px 0;
+  padding: 6px 6px;
+  margin: 5px 0;
   display: inline-block;
   border: 1px solid #ccc;
   border-radius: 4px;
@@ -285,8 +288,8 @@ input[type=submit] {
   width: 100%;
   background-color: #4CAF50;
   color: white;
-  padding: 10px 10px;
-  margin: 6px 0;
+  padding: 5px 5px;
+  margin: 3px 0;
   border: none;
   border-radius:4px;
   cursor: pointer;
