--- UI/test.html
@@ -1,157 +1,0 @@
-
-
-
-
- <!DOCTYPE html>
-<html>
-<head>
-  <title>AutoMart</title>
-  
-  <link href="css/project.css" rel="stylesheet" type="text/css">
-  <script type="text/javascript" src="js/first_page.js"></script>
-</head>
-<body>
-
-
-    <div class="div_container">
-      <div class="div_header">
-        <div class="logo">
-          <img src="image/index.jpg" alt="image logo">
-        </div>
-
-        <div class="title">
-          AutoMart
-        </div>
-
-        <div class="link">
-          
-          <ul>
-             <li><a class="active" href="index.html">Home</a></li>
-              <li><a href="account.html">Sign Up</a></li>
-              <li><a href="login.html">Login</a></li>
-          </ul>
-        </div>
-      </div>
-
-    <div class="div_meaning">  
-    <h1>Auto Mart</h1>
-        <p>
-           Auto Mart is an online marketplace for automobiles of diverse makes, model or body type. With
-           Auto Mart, users can sell their cars or buy from trusted dealerships or private sellers.
-        </p>     
-    </div>
-<div class="car_container">
-  <div class="preview">
-    <div class="preview_card">
-      <div id="new_car">
-        <img src="image/index5.jpg"  alt="Sample photo">
-        </div>
-          <div class="preview_text">
-            <h3>Details Information</h3>
-            <p>
-              <ul>
-                <li>created on:25-07-2018</li>
-                <li>Price:300$</li>
-                <li>State:New</li>
-                <li>status:available</li>
-                <li>Model:V8</li>
-                <li>Body Type:Tax</li>
-                <li>Manufacturer:Benze</li>
-              </ul>
-              
-            </p>
-            
-          </div>
-          </div>
-  </div>
-
-
-
-  <div class="car_list">
-     <h3>the latest car you can purchase</h3>
-      <div class="div_list">
-        <div class="car_card">
-        <img src="image/index.jpg"  alt="Sample photo">
-          <div class="text">
-            <h3>Car model</h3>
-            <p>
-              created on:25-07-2018<br><br>
-              Price:300$<br><br>
-              State:New
-            </p>
-            <button id="index" onclick="get_image(this.id)">more</button>
-          </div>
-          </div>
-
-          <div class="car_card">
-        <img src="image/index5.jpg"  alt="Sample photo">
-          <div class="text">
-               <h3>Car model</h3>
-            <p>
-              created on:25-07-2018<br><br>
-              Price:300$<br><br>
-              State:New
-            </p>
-            <button id="index5" onclick="get_image(this.id)" >more</button>
-          </div>
-          </div>
-
-
-          <div class="car_card">
-        <img src="image/index2.jpg"  alt="Sample photo">
-          <div class="text">
-             <h3>Car model</h3>
-            <p>
-              created on:25-07-2018<br><br>
-              Price:300$<br><br>
-              State:New
-            </p>
-            <button id="index2" onclick="get_image(this.id)">more</button>
-          </div>
-          </div>
-
-
-        <div class="car_card">
-        <img src="image/index3.jpg"  alt="Sample photo">
-          <div class="text">
-             <h3>Car model</h3>
-            <p>
-              created on:25-07-2018<br><br>
-              Price:300$<br><br>
-              State:New
-            </p>
-            <button id="index3" onclick="get_image(this.id)">more</button>
-          </div>
-          </div>
-
-          <div class="car_card">
-        <img src="image/index4.jpg" alt="Sample photo">
-          <div class="text">
-             <h3>Car model</h3>
-            <p>
-              created on:25-07-2018<br><br>
-              Price:300$<br><br>
-              State:New
-            </p>
-            <button id="index4" onclick="get_image(this.id)">more</button>
-          </div>
-          </div>
-
-        
-    </div>
-    </div>
-  </div>
-
-
-<!-- Footer -->
-<div class="footer">
-   <div class="div_rule">
-      <!-- this wiil dispaly some rule and regulation aboout the page and its use-->
-      with the AutoMart you can post a car sale advertisement also you can buy car, please work with us through making <a href="account.html">sign up</a> to AutoMart sytem
-
-    </div>
-
-</div>
-</div>
-</body>
-</html>
