--- UI/test1.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<title></title>
-
-<style>
-	display:flex;
-	flex-
-</style>
-</head>
-<body>
-
-<div class="cont">
-	<div>
-		fffffffffff
-	</div>
-
-	<div>
-		ggggggggggggg
-	</div>
-</div>
-
-
-
-
-
-
-</body>
-</html>
