+++ UI/users/buyer_newprice.html
@@ -32,7 +32,7 @@
    <div class="div_car" style="width: 100%;">
     <div class="row-flex" style=" display:flex;flex-direction:column;align-items: center;">
 
-      <div class="new_car" style="float:none; margin-left:0%; width:40%">
+      <div class="new_car" style="float:none; margin-left:0%; width:30%">
       
         
         <fieldset>
