+++ UI/users/car.html
@@ -15,7 +15,7 @@
 </head>
 <body onload="car_option1()">
 
-  <div class="div_container">
+<div class="div_container" style="background-color:lightgrey; background-image:none;color:black">
 
     <div class="div_header">
       <div class="logo">
