+++ UI/users/car_advert.html
@@ -33,7 +33,7 @@
     <div class="div_car" style="width: 100%;">
     <div class="row-flex" style=" display:flex;flex-direction:column;align-items: center;">
 
-      <div class="new_car" style="float:none; margin:3%; width:40%">
+      <div class="new_car" style="float:none; margin:3%; width:30%">
          <!-- when the user change manufacturer the all the car will atomatically displayed here -->
        to advertise car you can upload more image of car. click on <b>Add Photo</b> button and upload more image
        <br><br>
