+++ UI/users/purchase.html
@@ -12,6 +12,7 @@
       flex-wrap: wrap;
       background-color:white;
       width:60%;
+      color:black;
     }
     .car_image img{
       width:100%;
@@ -20,7 +21,7 @@
 </head>
 <body>
 
-  <div class="div_container">
+  <div class="div_container" style="background-color:lightgrey; background-image:none;color:black">
     <div class="div_header">
       <div class="logo">
         <img src="../image/index.jpg" alt="image logo">
