+++ UI/users/report.html
@@ -33,7 +33,7 @@
        <div class="div_car" style="width: 100%;">
         <div class="row-flex" style=" display:flex;flex-direction:column;align-items: center;">
 
-          <div class="new_car" style="float:none; margin-left:0%; width:40%">
+          <div class="new_car" style="float:none; margin-left:0%; width:30%">
         
         <fieldset>
                 <legend><h3>Report Fraudulent car advertisement here</h3></legend>
@@ -49,7 +49,7 @@
                       <input type="text" id="reason_car" name="reason" placeholder="Reporting reason......">
 
                       <label for="description">Description</label>
-                      <textarea cols="5" rows="5"  id="description" name="description" placeholder="Reporting Description......"></textarea>
+                      <textarea cols="3" rows="3"  id="description" name="description" placeholder="Reporting Description......"></textarea>
                      
                       <input type="submit" value="Report Car" onclick="return reporting_car()">
                   </form>
