+++ package.json
@@ -9,8 +9,8 @@
     "lint": "eslint 'src/**/*.js' --fix",
     "format": "prettier-eslint 'src/**/*.{js,json}' --write",
     "test": "nyc --reporter=html --reporter=text mocha --require @babel/register test/*.js --exit",
-    "coverage":"nyc --reporter=lcov npm run test",
-    "coveralls": "nyc report --reporter=text-lcov | coveralls"  
+    "coverage": "nyc --reporter=lcov npm run test",
+    "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
     "type": "git",
